requirements-prod.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.30.0
Authlib==1.5.1
black==24.4.2
build==1.2.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudinary==1.43.0
coverage==7.5.3
cryptography==44.0.2
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
factory-boy==3.3.0
Faker==25.3.0
fastapi==0.115.11
flake8==7.0.0
google-api-core==2.25.0rc0
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
iniconfig==2.0.0
ipython==8.25.0
isort==5.13.2
jedi==0.19.1
mailjet-rest==1.3.4
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pip-tools==7.4.1
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.45
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1
pyasn1_modules
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyproject_hooks==1.1.0
pytest==8.2.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.0.4
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.46.1
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.30.0
Authlib==1.5.1
black==24.4.2
build==1.2.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudinary==1.43.0
coverage==7.5.3
cryptography==44.0.2
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
factory-boy==3.3.0
Faker==25.3.0
fastapi==0.115.11
flake8==7.0.0
google-api-core==2.25.0rc0
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
iniconfig==2.0.0
ipython==8.25.0
isort==5.13.2
jedi==0.19.1
mailjet-rest==1.3.4
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pip-tools==7.4.1
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.45
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyproject_hooks==1.1.0
pytest==8.2.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.0.4
requests==2.32.3
resend==2.10.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.46.1
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
